add locked mutations

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -35,19 +35,21 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, tokenPhrase),
     };
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
         data: args.data,
       },
       info
     );
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    return prisma.mutation.deleteUser({ where: { id: userId } }, info);
   },
   createPost(parent, args, { prisma, request }, info) {
     const userId = getUserId(request);
@@ -67,34 +69,75 @@ const Mutation = {
       info
     );
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error("Unable to update post");
+    }
     return prisma.mutation.updatePost(
       { where: { id: args.id }, data: args.data },
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error("Unable to delete post");
+    }
     return prisma.mutation.deletePost({ where: { id: args.id } }, info);
   },
-  createComment(parent, args, { prisma }, info) {
+  createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
-          author: { connect: { id: args.data.author } },
+          author: { connect: { id: userId } },
           post: { connect: { id: args.data.post } },
         },
       },
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!commentExists) {
+      throw new Error("Unable to update comment");
+    }
     return prisma.mutation.updateComment(
       { where: { id: args.id }, data: args.data },
       info
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!commentExists) {
+      throw new Error("Unable to delete comment");
+    }
     return prisma.mutation.deleteComment({ where: { id: args.id } }, info);
   },
 };
